Fix bug that would appear to add user even if was already taken.

+++ src/persistence/AlumniPersistence.java
@@ -46,13 +46,17 @@ public class AlumniPersistence extends AlumNetPersistence{
                 }
             } catch (Exception e) {
                 //failed to create the student object... delete the login
+                if (e.getMessage().equals("com.microsoft.sqlserver.jdbc.SQLServerException: The conversion of a date data type to a datetime data type resulted in an out-of-range value.")) {
+                    throw e;
+                }
 
                 if (loginService.deleteLogin(temp).equals(ANConstants.SUCCESS)) {
                     System.out.println("Successfully cleaned up after failing to add");
+                    throw new Exception("Successfully cleaned up after failing to add");
                 } else {
                     System.out.println("Failed to clean up Login table... please contact administrator");
+                    throw new Exception("Failed to clean up Login table... please contact administrator");
                 }
-                throw new Exception(e.getMessage() + "... Failed to clean up Login table... please contact administrator");
             }
         } catch (Exception e) {
             throw e;
